Fully implement v1 Quad space time protocol with support for moving to protocol v2.0

+++ contracts/mocks/QuadTokenMock.sol
@@ -5,7 +5,7 @@ import "../tokens/QuadToken.sol";
 // mock class using BasicToken
 contract QuadTokenMock is QuadToken {
   function QuadTokenMock(  
-      uint64[] _tokenIds,
+      uint256[] _tokenIds,
       address owner
     ) 
     public
